integer*8 error

+++ src/rys_roots.c
@@ -1394,7 +1394,7 @@ static void R_dnode(double *a, double *rt, FINT k)
         if (prod < 0) {
             goto L20;
         }
-        fprintf(stderr, " 0ROOT NUMBER %4d WAS NOT FOUND FOR POLYNOMIAL OF ORDER %4d\n", m, k);
+        fprintf(stderr, " 0ROOT NUMBER %4d WAS NOT FOUND FOR POLYNOMIAL OF ORDER %4d\n", (int)m, (int)k);
 L20:
         r5 = r1;
         p5 = p1;
@@ -1528,7 +1528,7 @@ void gamma_inc_like(double *ff, double x, FINT n)
         if (nterm > 199) {
                 fprintf(stderr, "libcint::rys_roots power series of gamma_inc_like not converge"
                         "val=%.16g last term=%.16g x=%.16g n=%d\n",
-                        suma, term[199], x, n);
+                        suma, term[199], x, (int)n);
                 exit(1);
         }
 
@@ -1730,7 +1730,7 @@ static void R_qnode(long double *a, long double *rt, FINT k)
         if (prod < 0) {
             goto L20;
         }
-        fprintf(stderr, " 0ROOT NUMBER %4d WAS NOT FOUND FOR POLYNOMIAL OF ORDER %4d\n", m, k);
+        fprintf(stderr, " 0ROOT NUMBER %4d WAS NOT FOUND FOR POLYNOMIAL OF ORDER %4d\n", (int)m, (int)k);
 L20:
         r5 = r1;
         p5 = p1;
@@ -1858,7 +1858,7 @@ static void qgamma_inc_like(long double *ff, long double x, FINT n)
         if (nterm > 199) {
                 fprintf(stderr, "libcint::rys_roots power series of gamma_inc_like not converge"
                         "val=%.16Lg last term=%.16Lg x=%.16Lg n=%d\n",
-                        suma, term[199], x, n);
+                        suma, term[199], x, (int)n);
                 exit(1);
         }
